refactor: WebsocketRouter and HTTPRouter
