+++ src/router/router.ts
@@ -19,9 +19,7 @@ import {
 import { trimSlash } from './utils.ts';
 import { MiddlewareExecutor } from './middleware/executor.ts';
 import { NextFunction } from './middleware/decorator.ts';
-import { HonoRequest } from 'https://deno.land/x/hono@v4.0.8/request.ts';
-import { getPath } from 'https://deno.land/x/hono@v4.0.8/utils/url.ts';
-import { RegExpRouter, SmartRouter, TrieRouter } from 'https://deno.land/x/hono@v4.0.8/mod.ts';
+import { resolveMethodParam } from './controller/params/resolver.ts';
 
 // deno-lint-ignore no-explicit-any
 export type Callback = (...args: any[]) => unknown;
@@ -40,7 +38,7 @@ export type ExecutionContext = HttpContext & {
 	websocketTopic?: string;
 };
 
-export class DanetRouter {
+export class DanetHTTPRouter {
 	private logger: Logger = new Logger('Router');
 	private methodsMap: Map<string, HandlerInterface>;
 	public prefix?: string;
@@ -140,121 +138,7 @@ export class DanetRouter {
 		this.prefix = prefix;
 	}
 
-	registerControllers(Controllers: Constructor[]) {
-		Controllers.forEach((Controller) => {
-			const httpEndpoint = MetadataHelper.getMetadata<string>('endpoint', Controller);
-			const webSocketEndpoint = MetadataHelper.getMetadata<string>('websocket-endpoint', Controller);
-			if (webSocketEndpoint)
-				this.registerWebSocketController(Controller, webSocketEndpoint);
-			else
-				this.registerController(Controller, httpEndpoint)
-			
-		});
-	}
-
-	private registerWebSocketController(Controller: Constructor, endpoint: string) {
-		endpoint = trimSlash(endpoint);
-		const path = (endpoint ? ('/' + endpoint) : '');
-		const methods = Object.getOwnPropertyNames(Controller.prototype);
-		const smartRouter = new SmartRouter({
-			routers: [new RegExpRouter(), new TrieRouter()],
-		  });
-		for (const methodName of methods) {
-			if (
-				methodName === 'constructor' ||
-				(Object.values(hookName) as string[]).includes(methodName)
-			) {
-				continue;
-			}
-			const controllerMethod = Controller.prototype[methodName];
-			const topic = MetadataHelper.getMetadata<string>(
-				'websocket-topic',
-				controllerMethod,
-			);
-			smartRouter.add('POST', topic, methodName);
-		}
-		this.router.get(path,
-			async (context: HttpContext) => {
-				const { response, socket } = Deno.upgradeWebSocket(context.req.raw)
-				const _id = crypto.randomUUID();
-				(context as ExecutionContext)._id = _id;
-				(context as ExecutionContext).getClass = () => Controller;
-				context.set('_id', _id);
-				(context as ExecutionContext).websocket = socket;
-				const executionContext = context as unknown as ExecutionContext;
-				const controllerInstance = await this.injector.get(
-					Controller,
-					executionContext,
-					// deno-lint-ignore no-explicit-any
-				) as any;
-				socket.onopen = async () => {
-					try {
-						await this.guardExecutor.executeAllRelevantGuards(
-							executionContext,
-							Controller,
-							() => ({}),
-						);
-					} catch (e) {
-						socket.close(1008, 'Unauthorized');
-					}
-				};
-				socket.onmessage = async (event) => {
-					
-					const { topic, data } = JSON.parse(event.data);					
-					const fakeRequest = new Request(`https://fakerequest.com/${topic}`, { method: 'POST', body: JSON.stringify(data) });
-					const [methods, foundParam] = smartRouter.match('POST', topic);
-					const methodName = methods[0][0] as string;
-					const messageExecutionContext = {} as ExecutionContext;
-
-					messageExecutionContext.req = new HonoRequest(fakeRequest, getPath(fakeRequest), [methods, foundParam] as any);
-					const _id = crypto.randomUUID();
-					messageExecutionContext._id = _id;
-					messageExecutionContext.getClass = () => Controller;
-					messageExecutionContext.getHandler = () => controllerInstance[methodName];
-					messageExecutionContext.websocketTopic = topic;
-					messageExecutionContext.websocketMessage = data;
-					await this.middlewareExecutor.executeAllRelevantMiddlewares(
-						context as unknown as ExecutionContext,
-						Controller,
-						controllerInstance[methodName],
-						async () => {
-							try {
-								await this.guardExecutor.executeAllRelevantGuards(
-									messageExecutionContext,
-									(() => ({})) as any,
-									controllerInstance[methodName],
-								);
-							} catch (e) {
-								socket.close(1008, 'Unauthorized');
-								return;
-							}
-							const params = await this.resolveMethodParam(
-								Controller,
-								controllerInstance[methodName],
-								messageExecutionContext,
-							);
-							let response;
-							try {
-								response = await controllerInstance[methodName](...params);
-							} catch (error) {
-								response = await this.filterExecutor
-									.executeControllerAndMethodFilter(
-										executionContext,
-										error,
-										Controller,
-										controllerInstance[methodName],
-									);
-							}
-							if (response)
-								socket.send(JSON.stringify(response));
-						});
-				};
-				return response;
-			}
-		);
-	}
-
-	private registerController(Controller: Constructor, basePath: string) {
+	public registerController(Controller: Constructor, basePath: string) {
 		const methods = Object.getOwnPropertyNames(Controller.prototype);
 		this.logger.log(
 			`Registering ${Controller.name} ${basePath ? basePath : '/'}`,
@@ -282,7 +166,7 @@ export class DanetRouter {
 					Controller,
 					ControllerMethod,
 				);
-				const params = await this.resolveMethodParam(
+				const params = await resolveMethodParam(
 					Controller,
 					ControllerMethod,
 					executionContext,
@@ -378,23 +262,5 @@ export class DanetRouter {
 		return context.res;
 	}
 
-	private async resolveMethodParam(
-		Controller: ControllerConstructor,
-		// deno-lint-ignore no-explicit-any
-		ControllerMethod: (...args: any[]) => unknown,
-		context: ExecutionContext,
-	) {
-		const paramResolverMap: Map<number, Resolver> = MetadataHelper.getMetadata(
-			argumentResolverFunctionsMetadataKey,
-			Controller,
-			ControllerMethod.name,
-		);
-		const params: unknown[] = [];
-		if (paramResolverMap) {
-			for (const [key, value] of paramResolverMap) {
-				params[key] = await value(context);
-			}
-		}
-		return params;
-	}
+
 }
